fix NPE in delete for invoke wih exception

+++ com.oracle.max.graal.nodes/src/com/oracle/max/graal/nodes/InvokeWithExceptionNode.java
@@ -155,7 +155,7 @@ public class InvokeWithExceptionNode extends ControlSplitNode implements Node.It
     public void delete() {
         FrameState stateAfter = stateAfter();
         super.delete();
-        if (stateAfter.usages().isEmpty()) {
+        if (stateAfter != null && stateAfter.usages().isEmpty()) {
             stateAfter.delete();
         }
     }
